lc3: setup_decoder: allow downsampling from source
There appears to be a bug where downsampling from received data `sr_hz`
to pcm data `sr_pcm_hz` is disallowed for the decoder, which causes the
setup to fail.

The encoder, on the other hand, sources the data from pcm, and allows
downsampling from that to the other end.

Assuming symmetry w.r.t. the direction of downsampling, the comparison
operator should be flipped.